refactor(core): refactored uas template code for core
